fix: address reviewer comments on request events and broker
- Preserve request-scoped context in goroutine for logging/event publishing
- Change RequestEventPayload token fields from int to int64 to match UsageLog
- Remove stored context.Context from Subscriber struct (containedctx lint)
- Add test EventBroker helper for RequestService tests

+++ internal/server/events/types.go
@@ -38,15 +38,15 @@ type RequestEventPayload struct {
 	Stream    bool   `json:"stream"`
 
 	// Token usage data
-	PromptTokens     int `json:"prompt_tokens,omitempty"`
-	CompletionTokens int `json:"completion_tokens,omitempty"`
-	TotalTokens      int `json:"total_tokens,omitempty"`
+	PromptTokens     int64 `json:"prompt_tokens,omitempty"`
+	CompletionTokens int64 `json:"completion_tokens,omitempty"`
+	TotalTokens      int64 `json:"total_tokens,omitempty"`
 
 	// Cache tokens
-	PromptCachedTokens        int `json:"prompt_cached_tokens,omitempty"`
-	PromptWriteCachedTokens   int `json:"prompt_write_cached_tokens,omitempty"`
-	PromptWriteCachedTokens5m int `json:"prompt_write_cached_tokens_5m,omitempty"`
-	PromptWriteCachedTokens1h int `json:"prompt_write_cached_tokens_1h,omitempty"`
+	PromptCachedTokens        int64 `json:"prompt_cached_tokens,omitempty"`
+	PromptWriteCachedTokens   int64 `json:"prompt_write_cached_tokens,omitempty"`
+	PromptWriteCachedTokens5m int64 `json:"prompt_write_cached_tokens_5m,omitempty"`
+	PromptWriteCachedTokens1h int64 `json:"prompt_write_cached_tokens_1h,omitempty"`
 
 	// Include enough data for UI to update without additional fetch
 	APIKeyID  *int      `json:"api_key_id,omitempty"`
